Makes Captains Quarters follow solid principles
Took captains quarters and made it an inteface instead of implimenting it with the square class.
