+++ src/main/java/cs361/battleships/models/Board.java
@@ -52,7 +52,6 @@ public class Board {
 	}
 
 	private Result attack(Square s) {
-		/* Commented out to allow multiple hits in the same location*/
 //		if (attacks.stream().anyMatch(r -> r.getLocation().equals(s))) {
 //			var attackResult = new Result(s);
 //			attackResult.setResult(AtackStatus.INVALID);
