+++ src/main/java/cs361/battleships/models/Game.java
@@ -25,7 +25,7 @@ public class Game {
         do {
             // AI places random ships, so it might try and place overlapping ships
             // let it try until it gets it right
-            opponentPlacedSuccessfully = opponentsBoard.placeShip(ship, randRow(), randCol(), randVertical());
+            opponentPlacedSuccessfully = opponentsBoard.placeShip(ship, x, y, isVertical /*ship, randRow(), randCol(), randVertical()*/);
         } while (!opponentPlacedSuccessfully);
 
         return true;
